+++ src/Components/slate/PreferenceForm.js
@@ -4,9 +4,6 @@ import "./preferences.css"
 const PreferenceForm = ({ onSubmit }) => {
     return (
         <form className="prefForm" onSubmit={onSubmit}>
-            <label htmlFor="myname">Enter username</label>
-            <input id="myname" name="myname" type="text" /><br />
-
             <label htmlFor="sport">Enter your favorite sport</label>
             <input id="sport" name="sport" type="text" /><br />